requirements-benchmark-ocr.txt
jiwer
pandas
pillow
pymupdf
easyocr
pytesseract
transformers
torch
sentencepiece
accelerate
numpy
mistralai
